requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
email_validator==2.2.0
pillow==10.4.0
python-dotenv==1.0.1
gunicorn==22.0.0
